Refactor cells, constants, and tables
