+++ clients/admin-ui/src/features/custom-fields/constants.ts
@@ -1,0 +1,4 @@
+export const FIELD_TYPE_MAP = new Map([
+  ["string", "Single Select"],
+  ["string[]", "Multi Select"],
+]);
