+++ clients/admin-ui/src/features/privacy-notices/constants.ts
@@ -9,8 +9,3 @@ export const ENFORCEMENT_LEVEL_MAP = new Map([
   ["frontend", "Front end"],
   ["not_applicable", "Not applicable"],
 ]);
-
-export const FIELD_TYPE_MAP = new Map([
-  ["string", "Single Select"],
-  ["string[]", "Multi Select"],
-]);
